move buttons to package-button from ConverterScreen
