+++ app/src/main/java/ksnd/open/hiraganaconverter/view/screen/ConverterScreen.kt
@@ -1,40 +1,27 @@
 package ksnd.open.hiraganaconverter.view.screen
 
 import android.widget.Toast
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.animation.scaleIn
-import androidx.compose.animation.scaleOut
-import androidx.compose.animation.slideIn
-import androidx.compose.animation.slideOut
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.defaultMinSize
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.Divider
 import androidx.compose.material.OutlinedTextField
 import androidx.compose.material.TextFieldDefaults
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FilledTonalButton
-import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedCard
 import androidx.compose.material3.Scaffold
@@ -44,10 +31,7 @@ import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
@@ -64,15 +48,15 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
-import kotlinx.coroutines.launch
 import ksnd.open.hiraganaconverter.R
 import ksnd.open.hiraganaconverter.view.parts.ConversionTypeSpinnerCard
 import ksnd.open.hiraganaconverter.view.parts.TopBar
+import ksnd.open.hiraganaconverter.view.parts.button.ConvertButton
 import ksnd.open.hiraganaconverter.view.parts.button.CustomFilledTonalIconButton
+import ksnd.open.hiraganaconverter.view.parts.button.MoveTopButton
 import ksnd.open.hiraganaconverter.view.rememberButtonScaleState
 import ksnd.open.hiraganaconverter.viewmodel.ConvertViewModel
 import ksnd.open.hiraganaconverter.viewmodel.ConvertViewModelImpl
@@ -109,7 +93,7 @@ private fun ConverterScreenContent(viewModel: ConvertViewModel) {
     Scaffold(
         topBar = { TopBar(scrollBehavior) },
         containerColor = MaterialTheme.colorScheme.surface,
-        floatingActionButton = { TopButton(scrollState = scrollState) },
+        floatingActionButton = { MoveTopButton(scrollState = scrollState) },
     ) { padding ->
         Column(
             modifier = Modifier
@@ -128,13 +112,11 @@ private fun ConverterScreenContent(viewModel: ConvertViewModel) {
             Row(
                 verticalAlignment = Alignment.CenterVertically,
             ) {
-                // 変換タイプ（ひらがな、カタカナ）を選択するスピナー
                 Row(modifier = Modifier.weight(1f)) {
                     ConversionTypeSpinnerCard(
                         onSelectedChange = viewModel::changeHiraKanaType,
                     )
                 }
-                // 変換するときに押すボタン
                 ConvertButton(
                     onClick = {
                         viewModel.convert(context = context)
@@ -173,46 +155,6 @@ private fun ConverterScreenContent(viewModel: ConvertViewModel) {
             )
 
             Spacer(modifier = Modifier.height(120.dp))
-        }
-    }
-}
-
-@Composable
-private fun ConvertButton(onClick: () -> Unit) {
-    val buttonScaleState = rememberButtonScaleState()
-    FilledTonalButton(
-        modifier = Modifier
-            .padding(all = 8.dp)
-            .height(48.dp)
-            .scale(scale = buttonScaleState.animationScale.value),
-        onClick = onClick,
-        colors = ButtonDefaults.filledTonalButtonColors(
-            containerColor = MaterialTheme.colorScheme.primaryContainer,
-        ),
-        interactionSource = buttonScaleState.interactionSource,
-    ) {
-        Row(
-            modifier = Modifier
-                .wrapContentWidth()
-                .fillMaxHeight(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.Center,
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.ic_baseline_compare_arrows_24),
-                contentDescription = "convert",
-                colorFilter = ColorFilter.tint(
-                    MaterialTheme.colorScheme.onPrimaryContainer,
-                ),
-                contentScale = ContentScale.Fit,
-                modifier = Modifier.size(36.dp),
-            )
-            Text(
-                text = stringResource(id = R.string.conversion),
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onPrimaryContainer,
-                modifier = Modifier.padding(start = 8.dp),
-            )
         }
     }
 }
@@ -361,37 +303,6 @@ private fun AfterTextField(
             .defaultMinSize(minHeight = 120.dp)
             .fillMaxWidth(),
     )
-}
-
-@OptIn(ExperimentalAnimationApi::class)
-@Composable
-private fun TopButton(scrollState: ScrollState) {
-    val scrollScope = rememberCoroutineScope()
-    val offset = IntOffset(x = 100, y = 100)
-    val showVisibleTopBar by remember {
-        derivedStateOf { scrollState.value > 0 }
-    }
-
-    AnimatedVisibility(
-        visible = showVisibleTopBar,
-        enter = scaleIn() + slideIn(initialOffset = { offset }),
-        exit = scaleOut() + slideOut(targetOffset = { offset }),
-    ) {
-        FloatingActionButton(
-            onClick = {
-                scrollScope.launch {
-                    scrollState.animateScrollTo(0)
-                }
-            },
-        ) {
-            Text(
-                modifier = Modifier.padding(vertical = 4.dp, horizontal = 16.dp),
-                text = stringResource(id = R.string.top),
-                color = MaterialTheme.colorScheme.onPrimaryContainer,
-                style = MaterialTheme.typography.bodyLarge,
-            )
-        }
-    }
 }
 
 @Preview
